Use percent-encoded commas to support email clients like Thunderbird
A somewhat strange bug. The mailto: link worked when opening Thunderbird
from Firefox, but didn't work when attempting to open Thunderbird from
Chromium-based browsers like Brave and Chromium.

The issue was with the comma, which is a reserved character. The
convertHTMLToPercentEncoded function has been refactored to convert
commas to percent-encoded commas.

There's also now a link to a live page using this script.